ecommerce_/requirements.txt
Django==5.2
mysqlclient==2.2.7
Pillow==11.2.1
python-dotenv==1.0.1
django-crispy-forms==2.1
django-cleanup==8.1.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pyOpenSSL==23.3.0
django-cors-headers==4.3.1 